Finally fixed bug with AsyncProcessor and derivatives
Derivatives of AsyncProcessor (ParamProcessor, in particular) we crashing. The problem was that AsyncProcessor is created first and immediately starts the processing thread upon constrcution. Consequently, the virtual methods of the partially constructed classes were being called before their constructors had a chance to finish.
I suspect this wasn't showing up with PersistenArgsProcessor because it is a template class, and so it might not get constructed in the same was as a regular class.