mexObjectInterface working
Continued progress on creating management class for interfacing to c++ objects using mex.
mexObjectInterface handles object creation, and dynamic interfacing with methods AND properties of the object.

The point of mexObjectInterface (as opposed to mexInterface and mexObjectManager) is to make it easier to create matlab classes which wrap the interface.
mexObjectInterface provides dynamic info about wether a memeber is a function or variable, static or non-static, hidden or visible, and it's intented accessibility (public, protected, private).
The idea is that using that info, a class's dynamic props could be set upon construction.